cleaned bash files, cleaned console output

+++ scripts/connectFFProcesses.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+#Let the program know we're starting this script
 echo "BASH: starting process connection script"
 
 #While the process with the first pid entered exists
@@ -8,9 +9,11 @@ while kill -0 $1 2> /dev/null; do
     sleep .1
 done
 
-#TODO: Check of proc 2 exists, only kill it if it does
 
-#It doesn't exist anymore. Kill the second process
-kill $2
+#It doesn't exist anymore. Kill the second process if it exists
+if kill -0 $2 2> /dev/null; then kill $2; fi
+
+#Wait a little for the console output to catch up
+sleep .25
 
 echo "BASH: finished process connection script"
